requirements.txt
numpy>=1.26
pydicom>=2.4
pandas>=2.2
matplotlib>=3.8
scipy>=1.11
scikit-image>=0.22
scikit-learn>=1.3
PyYAML>=6.0
torch>=2.2
monai>=1.3          # Medical imaging utilities (MONAI)
opencv-python>=4.9  # CLAHE contrast enhancement
